refactor: Separated login and signup services
